a whole bunch of changes in the xml service descriptions regarding default values and the definition of all allowed values.
Plus a our very first '''own''' vendor extension

     ''coherence:X_withVendorDefines = "True|False"''

which defines if a value in

 * a allowedValueList and
 * a allowedValueRange

can be extended by a vendor.

This is currently used only in the description files used for service creation, but it could be inserted into the description files we send out to other devices too.

This changeset fixes #37 and #38.