Flickr seems to have changed the way original photos can be retrieved.
[http://flickr.com/services/api/misc.urls.html Flickr Photo Source URLs]

To access the original photos, we need to construct the url with an ''originalsecret'', which isn't returned in the response to our current request.

So for now we just take the default image (medium resolution?) that Flickr offers and try to sort out later what's needed to retrieve that ''originalsecret'' value.

This closes #73.